7204: Setup for testing new `arg get` command.
